move out write
